431583: Bugzilla relang problem with new setup
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=431583

+++ org.eclipse.mylyn.bugzilla.releng/modules/bugzilla/manifests/site.pp
@@ -105,6 +105,7 @@ define bugzilla::site (
 #
     exec { "delete bugzilla $version":
       command => "rm -R $base/$version",
+      onlyif    => "/usr/bin/test -d $base/$version",
       cwd     => "$base",
       user => "root",
       timeout => 300,
